backend/requirements.txt
flask
Pillow
torch
tensorflow
ultralytics
seaborn